Control.Pipe.C3

Pipes

commandSender :: (Serialize a, Serialize b) => a -> Handler (Maybe b)

Send a single command over the outgoing pipe and wait for a response. If the incoming pipe is closed before a response arrives, returns Nothing.

commandReceiver :: (Serialize a, Serialize b) => (a -> IO b) -> Handler ()

Wait for commands on the incoming pipe, handle them, and send the reponses over the outgoing pipe.